fix(cart): booking session could be use pinna store
