+++ frontend/src/css/BookingView.css
@@ -1227,3 +1227,29 @@
 .fc .fc-toolbar-title {
   font-size: 0.875rem !important;
 }
+
+/* Suggestion events - read-only appearance (no hover effect) */
+:deep(.fc-event.suggestion-exact),
+:deep(.fc-event.suggestion-alt) {
+  cursor: default !important;
+  opacity: 0.9;
+}
+
+:deep(.fc-event.suggestion-exact:hover),
+:deep(.fc-event.suggestion-alt:hover) {
+  opacity: 0.9 !important;
+  transform: none !important;
+  box-shadow: none !important;
+}
+
+/* Booked events - already confirmed bookings */
+:deep(.fc-event.booked-event) {
+  cursor: not-allowed !important;
+  opacity: 0.6;
+  pointer-events: none !important;
+}
+
+:deep(.fc-event.booked-event:hover) {
+  opacity: 0.6 !important;
+  transform: none !important;
+}
